+++ packages/angular/projects/angular-core/src/forms/textarea/textarea.directive.ts
@@ -1,5 +1,17 @@
-import {Directive, ElementRef, forwardRef, HostListener} from '@angular/core';
-import {ControlValueAccessor, NG_VALUE_ACCESSOR} from '@angular/forms';
+import {
+  Directive,
+  ElementRef,
+  forwardRef,
+  HostListener,
+  Inject,
+  Injector,
+  Renderer2
+} from '@angular/core';
+import {
+  NG_VALIDATORS,
+  NG_VALUE_ACCESSOR
+} from '@angular/forms';
+import { FormControlElementDirective } from '../form-control/form-control-element.directive';
 
 @Directive({
   // eslint-disable-next-line @angular-eslint/directive-selector
@@ -9,56 +21,44 @@ import {ControlValueAccessor, NG_VALUE_ACCESSOR} from '@angular/forms';
       provide: NG_VALUE_ACCESSOR,
       useExisting: forwardRef(() => TextareaDirective),
       multi: true
+    },
+    {
+      provide: NG_VALIDATORS,
+      useExisting: forwardRef(() => TextareaDirective),
+      multi: true
     }
   ]
 })
 
-export class TextareaDirective implements ControlValueAccessor {
-  // eslint-disable-next-line @typescript-eslint/no-empty-function, @typescript-eslint/no-explicit-any
-  onChange: (value: any) => void = () => {};
-  // eslint-disable-next-line @typescript-eslint/no-empty-function, @typescript-eslint/no-explicit-any
-  onTouched: () => any = () => {};
+export class TextareaDirective extends FormControlElementDirective {
 
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  private _value: any;
+  #value?: string;
 
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  get value(): any {
-    return this._value;
+  get value(): string {
+    return this.#value ?? '';
   }
 
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  set value(val: any) {
-    if (val !== this._value) {
-      this._value = val;
-      this.onChange(this._value);
-      this.onTouched();
-      this.elementRef.nativeElement.textarea.value = val;
-    }
+  set value(val: string) {
+     if (val !== this.#value) {
+      this.#value = val;
+      this.onChange(this.#value);
+      this.validatorOnChange();
+     }
   }
 
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  writeValue(value: any): void {
+  writeValue(value: string): void {
     if (value) {
       this.value = value;
+      this.elementRef.nativeElement.textarea.value = value;
     }
   }
 
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  registerOnChange(fn: any): void {
-    this.onChange = fn;
+  constructor(public override elementRef: ElementRef, private renderer: Renderer2, @Inject(Injector) injector: Injector) {
+    super(elementRef, injector);
   }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  registerOnTouched(fn: any): void {
-    this.onTouched = fn;
-  }
-
-  constructor(private elementRef: ElementRef) {}
 
   @HostListener('input', ['$event.target.value'])
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  listenForValueChange(value: any): void {
+  listenForValueChange(value: string): void {
     this.value = value;
   }
 }