+++ packages/components/grid/src/filter-column.ts
@@ -1,5 +1,7 @@
-import type { PropertyValues, TemplateResult } from 'lit';
+import type { TemplateResult } from 'lit';
+import type { DataSourceFilterFunction } from '@sl-design-system/shared';
 import { getNameByPath, getValueByPath } from '@sl-design-system/shared';
+import { localized, msg } from '@lit/localize';
 import { html } from 'lit';
 import { property, state } from 'lit/decorators.js';
 import { GridColumn } from './column.js';
@@ -12,9 +14,16 @@ export interface GridFilterOption {
   value?: unknown;
 }
 
-export class GridFilterColumn extends GridColumn {
+let nextUniqueId = 0;
+
+@localized()
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export class GridFilterColumn<T = any> extends GridColumn<T> {
   /** The internal options if none are provided. */
   @state() internalOptions?: GridFilterOption[];
+
+  /** The filter function if you want to do custom filtering. */
+  @state() filter?: DataSourceFilterFunction<T>;
 
   /**
    * The mode for the filter:
@@ -38,15 +47,8 @@ export class GridFilterColumn extends GridColumn {
   override connectedCallback(): void {
     super.connectedCallback();
 
+    this.id ||= `grid-filter-${nextUniqueId++}`;
     this.scopedElements = { ...this.scopedElements, 'sl-grid-filter': GridFilter };
-  }
-
-  override willUpdate(changes: PropertyValues<this>): void {
-    super.willUpdate(changes);
-
-    if (changes.has('grid')) {
-      this.itemsChanged();
-    }
   }
 
   override itemsChanged(): void {
@@ -54,18 +56,31 @@ export class GridFilterColumn extends GridColumn {
 
     if (this.mode !== 'text' && typeof this.options === 'undefined') {
       // No options were provided, so we'll create a list of options based on the column's values
-      this.internalOptions = this.grid?.items
+      this.internalOptions = this.grid?.dataSource?.items
         ?.reduce((acc, item) => {
-          const value = getValueByPath(item, this.path);
+          let value = getValueByPath(item, this.path),
+            label = value?.toString() ?? '';
+
+          if (value === null || value === undefined || (typeof value === 'string' && value.trim() === '')) {
+            label = msg('Blank');
+            value = '';
+          }
 
           if (value !== null && !acc.some(option => option.value === value)) {
-            acc.push({ label: value?.toString() ?? '', value });
+            acc.push({ label, value });
           }
 
           return acc;
         }, [] as GridFilterOption[])
         .sort((a, b) => a.label.localeCompare(b.label));
     }
+  }
+
+  override stateChanged(): void {
+    super.stateChanged();
+
+    const filter = this.grid?.dataSource?.filters.get(this.id);
+    this.value = filter ? filter.value : undefined;
   }
 
   override renderHeader(): TemplateResult {
@@ -75,6 +90,7 @@ export class GridFilterColumn extends GridColumn {
           .column=${this}
           .mode=${this.mode || 'select'}
           .options=${this.options ?? this.internalOptions}
+          .path=${this.path}
           .value=${this.value}
         >
           ${this.header ?? getNameByPath(this.path)}