+++ packages/components/shared/src/popover.ts
@@ -33,8 +33,9 @@ function getOffset(element: HTMLElement): number {
 
 function roundByDPR(num: number): number {
   const dpr = window.devicePixelRatio || 1;
+  const rounded = Math.round(num * dpr) / dpr;
 
-  return Math.round(num * dpr) / dpr || -10000;
+  return !isNaN(rounded) ? rounded : -10000;
 }
 
 export const isPopoverOpen = (element?: HTMLElement): boolean => {