+++ packages/components/label/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @sl-design-system/label
 
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @sl-design-system/button@0.0.10
+  - @sl-design-system/text-input@0.1.5
+  - @sl-design-system/textarea@0.1.5
+
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @sl-design-system/button@0.0.9
+  - @sl-design-system/text-input@0.1.4
+  - @sl-design-system/textarea@0.1.4
+
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @sl-design-system/button@0.0.8
+  - @sl-design-system/text-input@0.1.3
+  - @sl-design-system/textarea@0.1.3
+
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @sl-design-system/button@0.0.7
+  - @sl-design-system/text-input@0.1.2
+  - @sl-design-system/textarea@0.1.2
+
 ## 0.1.2
 
 ### Patch Changes